requirements.txt
streamlit
torch
ultralytics
pillow
numpy
opencv-python-headless